requirements.txt
python-dotenv
streamlit
streamlit-extras
youtube-transcript-api>=0.6.0
google-generativeai
langcodes
language_data
